requirements.txt

streamlit==1.32.0
seaborn
scikit-learn
pandas==1.5.0
numpy
matplotlib
mysql-connector-python==8.0.27
joblib